# public/docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'CI4 Board API'
  description: "CodeIgniter 4 기반 헤드리스 게시판 REST API\n\n## 인증\n로그인 후 발급된 `access_token`을 `Authorization: Bearer {token}` 헤더로 전달합니다.\n\n## 기본 응답 형식\n```json\n{ \"success\": true,  \"data\": {},   \"message\": \"\" }\n{ \"success\": false, \"data\": null, \"message\": \"에러 메시지\" }\n```\n\n## API 구조\n| 경로 접두사 | 설명 | 인증 |\n|------------|------|------|\n| `/api/v1/*` | 사용자·게시판·쪽지·CMS | JWT (일부 공개) |\n| `/api/admin/v1/*` | 관리자 전용 | Admin JWT (role: superadmin/manager) |"
  version: 2.0.0
servers:
  -
    url: 'http://localhost:8080'
    description: '개발 서버'
paths:
...
  /api/admin/v1/cms/popups:
    get:
      tags:
        - AdminCMS
      summary: '팝업 목록'
      operationId: 196fe23b6f8eddb9a3110b20d801137e
      parameters:
        -
          name: page
          in: query
          schema:
            type: integer
            default: 1
        -
          name: is_used
          in: query
          schema:
            type: integer
            enum:
              - 0
              - 1
      responses:
        '200':
          description: '팝업 목록'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          BearerAuth: []
    post:
      tags:
        - AdminCMS
      summary: '팝업 생성'
      operationId: 92c549fa8ad01d8df5b5f0295d42b740
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - title
                - contents
              properties:
                title:
                  type: string
                contents:
                  type: string
                position:
                  type: string
                start_at:
                  type: integer
                end_at:
                  type: integer
                is_used:
                  type: boolean
                  default: true
              type: object
      responses:
        '201':
          description: '생성 완료'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          BearerAuth: []
  '/api/admin/v1/cms/popups/{idx}':
    put:
      tags:
        - AdminCMS
      summary: '팝업 수정'
      operationId: fd802d364033566e24051b226e1f6f37
      parameters:
        -
          name: idx
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                contents:
                  type: string
                position:
                  type: string
                start_at:
                  type: integer
                end_at:
                  type: integer
                is_used:
                  type: boolean
              type: object
      responses:
        '200':
          description: '수정 완료'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          BearerAuth: []
    delete:
      tags:
        - AdminCMS
      summary: '팝업 삭제'
      operationId: 5a48e97bf538b2024191ce02d4f1d817
      parameters:
        -
          name: idx
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: '삭제 완료'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          BearerAuth: []
...
  /api/admin/v1/members:
    get:
      tags:
        - AdminMember
      summary: '회원 목록'
      operationId: 301e870d45f2fc5e6329fecb463edd22
      parameters:
        -
          name: keyword
          in: query
          schema:
            type: string
        -
          name: status
          in: query
          schema:
            type: integer
            enum:
              - 0
              - 1
        -
          name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: '회원 목록 (pagination 포함)'
      security:
        -
          BearerAuth: []
  '/api/admin/v1/members/{idx}':
    put:
      tags:
        - AdminMember
      summary: '회원 정보 수정'
      operationId: 45d862efe2aae1e9daacdf316f5b7c6e
      parameters:
        -
          name: idx
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                group_idx:
                  type: integer
                status:
                  type: integer
                  enum: [0, 1]
                nickname:
                  type: string
                email:
                  type: string
                new_password:
                  type: string
                  minLength: 6
              type: object
      responses:
        '200':
          description: '수정 완료'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
      security:
        -
          BearerAuth: []
...
components:
  schemas:
    ApiResponse:
      properties:
        success:
          type: boolean
        data:
          nullable: true
        message:
          type: string
          nullable: true
      type: object
...
  responses:
    NotFound:
      description: '리소스를 찾을 수 없습니다.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Unauthorized:
      description: '인증이 필요합니다.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
...
    ValidationError:
      description: '유효성 검사 오류입니다.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
  securitySchemes:
    BearerAuth:
      type: http
      description: 'JWT Bearer token'
      bearerFormat: JWT
      scheme: bearer
...